Support for local (directory) requirements
For NPM install to work (under 3.10.10, which is what AWS Lambda uses), the "name" field is required in the package.json

I've added this as the directory name of the package, which might not be true (i.e. you could have a local package named `super-log` and drop it in a directory called `super-log-hack`). 

Unfortunately, I haven't delved deep enough into Lerna to figure out how to access the original package.json name.

I will test this myself now and comment if this solves my problem.